fix: bot WhatsApp LID migration - resolucion de LIDs con cache + conexion descartable
- Root cause: WhatsApp migro enrutamiento de contactos a @lid (2026-08-10).
  Enviar al JID numerico (@s.whatsapp.net) resuelve pero NO entrega (perdida silenciosa).

bot-furi/bot.js:
- Nueva seccion de resolucion de LIDs: lidCache, cargarLidsCache(),
  guardarLidCache(), lidJid(), resolverLidsSolo() (conexion descartable)
- main(): si faltan LIDs, los resuelve en conexion descartable ANTES de la principal
- enviarMensaje(): usa SOLO cache de LIDs (nunca llama onWhatsApp en principal)
- esperarAck(): flush periodico del event buffer cada 2s para liberar ACKs retenidos

bot-furi/.gitignore: +lids.json +*.txt
docs: bot-whatsapp.md, historial.md, errores-conocidos.md actualizados

Verificado: mood-120 -> ACK status=4 (leido) -> marcado en bot_notificaciones.
CI manual paso exitosamente.

+++ supabase_schema.sql
@@ -403,7 +403,7 @@ CREATE TABLE IF NOT EXISTS class_schedules (
   end_time TEXT DEFAULT '',
   professor TEXT DEFAULT '',
   user_id TEXT DEFAULT '',
-  color INTEGER DEFAULT 4286262670,
+  color BIGINT DEFAULT 4286262670,
   created_at TIMESTAMPTZ DEFAULT NOW(),
   updated_at TIMESTAMPTZ DEFAULT NOW()
 );
